feat: move Trace out of AppContainer
