+++ packages/platform-web/src/shuvi-app/react/view/ReactView.client.tsx
@@ -8,6 +8,7 @@ import Loadable from '../loadable';
 import { IReactClientView } from '../types';
 import { doRender } from './render';
 import { clientEntryTrace } from '../../entry/client/trace';
+import { Trace } from '../Trace';
 
 const headManager = new HeadManager();
 
@@ -67,7 +68,9 @@ export class ReactClientView implements IReactClientView {
       <Router router={router}>
         <AppContainer app={app}>
           <HeadManagerContext.Provider value={headManager.updateHead}>
-            <TypedAppComponent />
+            <Trace>
+              <TypedAppComponent />
+            </Trace>
           </HeadManagerContext.Provider>
         </AppContainer>
       </Router>
